requirements.txt
pandas>=2.1
openpyxl>=3.1
duckdb>=1.0
streamlit>=1.30
plotly>=5.20
python-dotenv>=1.0
pytest>=8.0
rich>=13.0

